Seperate IKE payload in message into an independent file
