+++ message/Payload_securityassociation.go
@@ -1,0 +1,213 @@
+package message
+
+import (
+	"encoding/binary"
+
+	"github.com/pkg/errors"
+)
+
+var _ IKEPayload = &SecurityAssociation{}
+
+type SecurityAssociation struct {
+	Proposals ProposalContainer
+}
+
+type ProposalContainer []*Proposal
+
+type Proposal struct {
+	ProposalNumber          uint8
+	ProtocolID              uint8
+	SPI                     []byte
+	EncryptionAlgorithm     TransformContainer
+	PseudorandomFunction    TransformContainer
+	IntegrityAlgorithm      TransformContainer
+	DiffieHellmanGroup      TransformContainer
+	ExtendedSequenceNumbers TransformContainer
+}
+
+type TransformContainer []*Transform
+
+type Transform struct {
+	TransformType                uint8
+	TransformID                  uint16
+	AttributePresent             bool
+	AttributeFormat              uint8
+	AttributeType                uint16
+	AttributeValue               uint16
+	VariableLengthAttributeValue []byte
+}
+
+func (securityAssociation *SecurityAssociation) Type() IKEPayloadType { return TypeSA }
+
+func (securityAssociation *SecurityAssociation) marshal() ([]byte, error) {
+	securityAssociationData := make([]byte, 0)
+
+	for proposalIndex, proposal := range securityAssociation.Proposals {
+		proposalData := make([]byte, 8)
+
+		if (proposalIndex + 1) < len(securityAssociation.Proposals) {
+			proposalData[0] = 2
+		} else {
+			proposalData[0] = 0
+		}
+
+		proposalData[4] = proposal.ProposalNumber
+		proposalData[5] = proposal.ProtocolID
+
+		proposalData[6] = uint8(len(proposal.SPI))
+		if len(proposal.SPI) > 0 {
+			proposalData = append(proposalData, proposal.SPI...)
+		}
+
+		// combine all transforms
+		var transformList []*Transform
+		transformList = append(transformList, proposal.EncryptionAlgorithm...)
+		transformList = append(transformList, proposal.PseudorandomFunction...)
+		transformList = append(transformList, proposal.IntegrityAlgorithm...)
+		transformList = append(transformList, proposal.DiffieHellmanGroup...)
+		transformList = append(transformList, proposal.ExtendedSequenceNumbers...)
+
+		if len(transformList) == 0 {
+			return nil, errors.Errorf("One proposal has no any transform")
+		}
+		proposalData[7] = uint8(len(transformList))
+
+		proposalTransformData := make([]byte, 0)
+
+		for transformIndex, transform := range transformList {
+			transformData := make([]byte, 8)
+
+			if (transformIndex + 1) < len(transformList) {
+				transformData[0] = 3
+			} else {
+				transformData[0] = 0
+			}
+
+			transformData[4] = transform.TransformType
+			binary.BigEndian.PutUint16(transformData[6:8], transform.TransformID)
+
+			if transform.AttributePresent {
+				attributeData := make([]byte, 4)
+
+				if transform.AttributeFormat == 0 {
+					// TLV
+					if len(transform.VariableLengthAttributeValue) == 0 {
+						return nil, errors.Errorf("Attribute of one transform not specified")
+					}
+					attributeFormatAndType := ((uint16(transform.AttributeFormat) & 0x1) << 15) | transform.AttributeType
+					binary.BigEndian.PutUint16(attributeData[0:2], attributeFormatAndType)
+					binary.BigEndian.PutUint16(attributeData[2:4], uint16(len(transform.VariableLengthAttributeValue)))
+					attributeData = append(attributeData, transform.VariableLengthAttributeValue...)
+				} else {
+					// TV
+					attributeFormatAndType := ((uint16(transform.AttributeFormat) & 0x1) << 15) | transform.AttributeType
+					binary.BigEndian.PutUint16(attributeData[0:2], attributeFormatAndType)
+					binary.BigEndian.PutUint16(attributeData[2:4], transform.AttributeValue)
+				}
+
+				transformData = append(transformData, attributeData...)
+			}
+
+			binary.BigEndian.PutUint16(transformData[2:4], uint16(len(transformData)))
+
+			proposalTransformData = append(proposalTransformData, transformData...)
+		}
+
+		proposalData = append(proposalData, proposalTransformData...)
+		binary.BigEndian.PutUint16(proposalData[2:4], uint16(len(proposalData)))
+
+		securityAssociationData = append(securityAssociationData, proposalData...)
+	}
+
+	return securityAssociationData, nil
+}
+
+func (securityAssociation *SecurityAssociation) unmarshal(b []byte) error {
+	for len(b) > 0 {
+		// bounds checking
+		if len(b) < 8 {
+			return errors.Errorf("Proposal: No sufficient bytes to decode next proposal")
+		}
+		proposalLength := binary.BigEndian.Uint16(b[2:4])
+		if proposalLength < 8 {
+			return errors.Errorf("Proposal: Illegal payload length %d < header length 8", proposalLength)
+		}
+		if len(b) < int(proposalLength) {
+			return errors.Errorf("Proposal: The length of received message not matchs the length specified in header")
+		}
+
+		proposal := new(Proposal)
+		var transformData []byte
+
+		proposal.ProposalNumber = b[4]
+		proposal.ProtocolID = b[5]
+
+		spiSize := b[6]
+		if spiSize > 0 {
+			// bounds checking
+			if len(b) < int(8+spiSize) {
+				return errors.Errorf("Proposal: No sufficient bytes for unmarshalling SPI of proposal")
+			}
+			proposal.SPI = append(proposal.SPI, b[8:8+spiSize]...)
+		}
+
+		transformData = b[8+spiSize : proposalLength]
+
+		for len(transformData) > 0 {
+			// bounds checking
+			if len(transformData) < 8 {
+				return errors.Errorf("Transform: No sufficient bytes to decode next transform")
+			}
+			transformLength := binary.BigEndian.Uint16(transformData[2:4])
+			if transformLength < 8 {
+				return errors.Errorf("Transform: Illegal payload length %d < header length 8", transformLength)
+			}
+			if len(transformData) < int(transformLength) {
+				return errors.Errorf("Transform: The length of received message not matchs the length specified in header")
+			}
+
+			transform := new(Transform)
+
+			transform.TransformType = transformData[4]
+			transform.TransformID = binary.BigEndian.Uint16(transformData[6:8])
+			if transformLength > 8 {
+				transform.AttributePresent = true
+				transform.AttributeFormat = ((transformData[8] & 0x80) >> 7)
+				transform.AttributeType = binary.BigEndian.Uint16(transformData[8:10]) & 0x7f
+
+				if transform.AttributeFormat == 0 {
+					attributeLength := binary.BigEndian.Uint16(transformData[10:12])
+					// bounds checking
+					if (12 + attributeLength) != transformLength {
+						return errors.Errorf("Illegal attribute length %d not satisfies the transform length %d",
+							attributeLength, transformLength)
+					}
+					copy(transform.VariableLengthAttributeValue, transformData[12:12+attributeLength])
+				} else {
+					transform.AttributeValue = binary.BigEndian.Uint16(transformData[10:12])
+				}
+			}
+
+			switch transform.TransformType {
+			case TypeEncryptionAlgorithm:
+				proposal.EncryptionAlgorithm = append(proposal.EncryptionAlgorithm, transform)
+			case TypePseudorandomFunction:
+				proposal.PseudorandomFunction = append(proposal.PseudorandomFunction, transform)
+			case TypeIntegrityAlgorithm:
+				proposal.IntegrityAlgorithm = append(proposal.IntegrityAlgorithm, transform)
+			case TypeDiffieHellmanGroup:
+				proposal.DiffieHellmanGroup = append(proposal.DiffieHellmanGroup, transform)
+			case TypeExtendedSequenceNumbers:
+				proposal.ExtendedSequenceNumbers = append(proposal.ExtendedSequenceNumbers, transform)
+			}
+
+			transformData = transformData[transformLength:]
+		}
+
+		securityAssociation.Proposals = append(securityAssociation.Proposals, proposal)
+
+		b = b[proposalLength:]
+	}
+
+	return nil
+}
